Refactor UI to reuse components and providers
