+++ src/main.tsx
@@ -1,7 +1,8 @@
 import { render } from 'preact'
-import App from './App'
+import App from './components/App.tsx'
 import Preferences from './components/Preferences'
 import './App.css'
+import { ConfigProvider } from './ConfigProvider'
 
 const urlParams = new URLSearchParams(window.location.search)
 const type = urlParams.get('type')
@@ -14,4 +15,9 @@ if (!rootElement) {
   throw new Error("Couldn't find root element")
 }
 
-render(<Component />, rootElement)
+render(
+  <ConfigProvider>
+    <Component />
+  </ConfigProvider>,
+  rootElement
+)
