+++ src/types.ts
@@ -1,0 +1,6 @@
+export type Config = {
+  theme: string
+  language: string
+  download_folder: string
+  open_on_download: string
+}
